fix: show the secondary status icon on switch toggle
